--- a/apps/test-app-runtime/package.json
+++ b/apps/test-app-runtime/package.json
@@ -9,11 +9,11 @@
     "eventual": "eventual"
   },
   "dependencies": {
-    "@aws-sdk/client-dynamodb": "^3.254.0",
-    "@aws-sdk/client-lambda": "^3.254.0",
-    "@aws-sdk/client-s3": "^3.254.0",
-    "@aws-sdk/client-sqs": "^3.254.0",
-    "@aws-sdk/lib-dynamodb": "^3.254.0",
+    "@aws-sdk/client-dynamodb": "^3.332.0",
+    "@aws-sdk/client-lambda": "^3.332.0",
+    "@aws-sdk/client-s3": "^3.332.0",
+    "@aws-sdk/client-sqs": "^3.332.0",
+    "@aws-sdk/lib-dynamodb": "^3.332.0",
     "@eventual/aws-client": "workspace:^",
     "@eventual/core": "workspace:^",
     "@eventual/integrations-slack": "workspace:^",
@@ -22,8 +22,8 @@
   },
   "devDependencies": {
     "@eventual/cli": "workspace:^",
-    "@types/aws-lambda": "^8.10.108",
-    "@types/jest": "^29",
+    "@types/aws-lambda": "^8.10.115",
+    "@types/jest": "^29.5.1",
     "@types/ms": "^0.7.31",
     "@types/node": "^18",
     "aws-embedded-metrics": "^4.1.0",