--- a/apps/tests/aws-runtime/package.json
+++ b/apps/tests/aws-runtime/package.json
@@ -14,21 +14,21 @@
     "test:runtime-deploy": "pnpm -r --filter tests-runtime run deploy:runtime && pnpm -r --filter tests-runtime run test:runtime"
   },
   "dependencies": {
-    "@aws-sdk/client-dynamodb": "^3.254.0",
-    "@aws-sdk/client-sqs": "^3.254.0",
-    "@aws-sdk/client-ssm": "^3.254.0",
+    "@aws-sdk/client-dynamodb": "^3.332.0",
+    "@aws-sdk/client-sqs": "^3.332.0",
+    "@aws-sdk/client-ssm": "^3.332.0",
     "@eventual/aws-client": "workspace:^",
     "@eventual/cli": "workspace:^",
     "@eventual/client": "workspace:^",
     "@eventual/core": "workspace:^",
-    "zod": "^3"
+    "zod": "^3.21.4"
   },
   "devDependencies": {
     "@anatine/zod-openapi": "^1.12.0",
-    "@aws-sdk/types": "^3.310.0",
-    "@jest/globals": "^29",
-    "@types/aws-lambda": "^8.10.108",
-    "@types/jest": "^29",
+    "@aws-sdk/types": "^3.329.0",
+    "@jest/globals": "^29.5.0",
+    "@types/aws-lambda": "^8.10.115",
+    "@types/jest": "^29.5.1",
     "@types/node": "^18",
     "esbuild": "^0.17.4",
     "jest": "^29",