--- a/apps/tests/aws-runtime/test/async-writer-handler.ts
+++ b/apps/tests/aws-runtime/test/async-writer-handler.ts
@@ -1,4 +1,4 @@
-import { Handler } from "aws-lambda";
+import type { Handler } from "aws-lambda";
 import { AWSHttpEventualClient } from "@eventual/aws-client";
 
 const serviceClient = new AWSHttpEventualClient({