--- a/package.json
+++ b/package.json
@@ -27,7 +27,7 @@
     "export": "turbo run export"
   },
   "devDependencies": {
-    "@types/jest": "^29.2.2",
+    "@types/jest": "^29.5.1",
     "@types/node": "^18",
     "@typescript-eslint/eslint-plugin": "^5.47.1",
     "@typescript-eslint/parser": "^5.47.1",
@@ -39,7 +39,7 @@
     "eslint-plugin-promise": "^6.1.1",
     "eslint-plugin-react": "^7.32.2",
     "husky": "^8.0.2",
-    "jest": "^29.3.1",
+    "jest": "^29.5.0",
     "lerna": "^6.6.2",
     "lint-staged": "^13.0.3",
     "prettier": "^2.8.8",