--- a/packages/@eventual/aws-runtime/package.json
+++ b/packages/@eventual/aws-runtime/package.json
@@ -26,15 +26,14 @@
     "@eventual/core": "workspace:^",
     "@eventual/core-runtime": "workspace:^",
     "aws-embedded-metrics": "^4.1.0",
-    "aws-lambda": "^1.0.7",
     "ulidx": "^0.3.0",
     "zod": "^3"
   },
   "devDependencies": {
     "@aws-sdk/types": "^3.254.0",
-    "@types/aws-lambda": "8.10.108",
-    "@types/express": "^4.17.15",
-    "@types/jest": "^29",
+    "@types/aws-lambda": "8.10.115",
+    "@types/express": "^4.17.17",
+    "@types/jest": "^29.5.1",
     "@types/node": "^18",
     "jest": "^29",
     "ts-jest": "^29",