--- a/packages/@eventual/compiler/package.json
+++ b/packages/@eventual/compiler/package.json
@@ -39,7 +39,7 @@
     "zod": "^3"
   },
   "devDependencies": {
-    "@types/jest": "^29",
+    "@types/jest": "^29.5.1",
     "@types/minimatch": "5.1.2",
     "@types/node": "^18",
     "esbuild": "^0.17.4",