--- a/packages/@eventual/core-runtime/package.json
+++ b/packages/@eventual/core-runtime/package.json
@@ -23,8 +23,8 @@
     "zod": "^3"
   },
   "devDependencies": {
-    "@jest/globals": "^29.3.1",
-    "@types/jest": "^29",
+    "@jest/globals": "^29.5.0",
+    "@types/jest": "^29.5.1",
     "@types/node": "^18",
     "jest": "^29",
     "ts-jest": "^29",