--- a/packages/@eventual/core/package.json
+++ b/packages/@eventual/core/package.json
@@ -27,9 +27,9 @@
     "zod": "^3"
   },
   "devDependencies": {
-    "@jest/globals": "^29.3.1",
+    "@jest/globals": "^29.5.0",
     "@tshttp/status": "^2.0.0",
-    "@types/jest": "^29",
+    "@types/jest": "^29.5.1",
     "@types/node": "^18",
     "jest": "^29",
     "ts-jest": "^29",