--- a/packages/@eventual/integrations-slack/package.json
+++ b/packages/@eventual/integrations-slack/package.json
@@ -24,7 +24,7 @@
   },
   "devDependencies": {
     "@eventual/core": "workspace:^",
-    "@types/jest": "^29",
+    "@types/jest": "^29.5.1",
     "@types/node": "^18",
     "@types/tsscmp": "^1.0.0",
     "itty-router": "2.6.6",