--- a/packages/@eventual/testing/package.json
+++ b/packages/@eventual/testing/package.json
@@ -20,8 +20,8 @@
   },
   "devDependencies": {
     "@aws-sdk/client-sqs": "^3.254.0",
-    "@jest/globals": "^29.3.1",
-    "@types/jest": "^29",
+    "@jest/globals": "^29.5.0",
+    "@types/jest": "^29.5.1",
     "@types/node": "^18",
     "jest": "^29",
     "ts-jest": "^29",