--- a/packages/@eventual/timeline/package.json
+++ b/packages/@eventual/timeline/package.json
@@ -20,8 +20,8 @@
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@types/react": "^18.0.26",
-    "@types/react-dom": "^18.0.9",
+    "@types/react": "^18.2.6",
+    "@types/react-dom": "^18.2.4",
     "@vitejs/plugin-react": "^3.0.0",
     "express": "^4.18.2",
     "get-port": "^6.1.2",