… packages/@eventual/aws-cdk/package.json
… packages/@eventual/cli/package.json
… pnpm-lock.yaml